src/core/requirements-external.txt
jsonschema>=3.2.0
jinja2>=2.11
pyyaml>=5.3.1
python-dotenv==0.13.*
tabulate==0.8.*
setuptools>=45.0.0
riotkit.pbs >= 1.0
virtualenv >= 20
gitignore_parser >= 0.0.8, < 0.1
